Add files via upload
Put a recycler view in dilog_view_folder to make the window bigger since it was so small before, I used the inputType multiline for the answer so the text would go to the next line instead of going off-screen. changed the layout of the add flashcards by making it a bit bigger with a space layout.


Things still needed; the flashcard save button needs to be implemented in order to save the actual flashcard, along with the delete function. Also for the design the name of the folders need to be bigger and need to figure out the background colors, the colors of the folders, flashcards as well